+++ app/models/conversion/voluntary/tasks/check_baseline.rb
@@ -1,3 +1,0 @@
-class Conversion::Voluntary::Tasks::CheckBaseline < TaskList::Task
-  attribute :confirm
-end